OrgRelLkp

Generates an organisational lookup table from the NHS Digital ODS ORD API based on organisational relationship data.

OrgRelLkp(PrimaryRole, NonPrimaryRole, RelTypes, RelPrimaryRoles, FromDate)

Arguments

PrimaryRole
The Primary Role code for organisations to be included in the lookup; quoted string, no default

NonPrimaryRole
The Non Primary Role code for organisations to be included in the lookup; quoted string, no default

RelTypes
The Named Relationship Types for related organisations to be included in the lookup; character vector, no default

RelPrimaryRoles
The Primary Role Ids for related organisations to be included in the lookup; character vector, no default

FromDate
The effective date from which to include Organisations operational on or after; character string in the format "yyyy-mm-dd", no default

Value

returns a data.frame of Organisation codes, names, start dates and end dates plus the organisations (code & name) they are related to and associated start and end dates for the relationship.

Notes

This function is experimental and has only been thoroughly QAd for use with the arguments presented in the first example in order to generate GP Practice to CCG lookup tables.

View the NHS Digital ODS API Implementation Guide at https://developer.nhs.uk/library/identifiers/ods-ord-api-implementation-guide/

View the NHS Digital ODS API Suite at https://directory.spineservices.nhs.uk/ODSAPISuite

See also

Examples

# return GP Practice to CCG Lookup to include all organisations effective on or after 01-04-2013

# NOT RUN {
OrgRelLkp(PrimaryRole    = "RO177",
          NonPrimaryRole = "RO76",
          RelTypes       = "RE4",
          RelPrimaryRoles = "RO98",
          FromDate       = "2013-04-01")
# }

# return GP practice to CCG/PCT Lookup to include all organisations effective on or after 01-04-2012

# NOT RUN {
OrgRelLkp("RO177","RO76","RE4",c("RO98","RO179"),"2012-04-01")
# }