getODS

Extracts summary ODS data for multiple organisations from the NHS Digital ODS ORD API into a data frame.

getODS(Name = "All", PostCode = "All", LastChangeDate = "All",
  Status = "All", PrimaryRoleId = "All", NonPrimaryRoleId = "All",
  OrgRecordClass = "All")

Arguments

Name
Search organisations based on name. Organisations that contain the argument string in their name are returned.; quoted string; default "All" applies no filter

PostCode
Search organisations based on postcode. Organisations that contain the argument string in their postcode are returned.; quoted string; default "All" applies no filter

LastChangeDate
Search for organisations based on their last changed date. Date must be in format "YYYY-MM-DD". The search is greater than or equal to. Dates are restricted to 185 days from present.; quoted string; default "All" applies no filter

Status
Search for organisations based on their active status. Arguments can be "Active" or "Inactive".; quoted string; default "All" applies no filter

PrimaryRoleId
Search for organisations based on their primary role codes.; quoted string; default "All" applies no filter

NonPrimaryRoleId
Search for organisations based on their non primary role codes.; quoted string; default "All" applies no filter

OrgRecordClass
Search for organisations based on their record class. Arguments can be "RC1" or "RC2".; quoted string; default "All" applies no filter

Value

returns a data.frame containing the following details for the organisations that meet the filter specifications: Name, Organisation ID, Status, Organisation Record Class, Postcode, Last Change Date, Primary Role ID, Non primary Role ID, Primary Role Description, Organisation Link (API endpoint URL for full organisation record)

Notes

View the NHS Digital ODS API Implementation Guide at https://developer.nhs.uk/library/identifiers/ods-ord-api-implementation-guide/

View the NHS Digital ODS API Suite at https://directory.spineservices.nhs.uk/ODSAPISuite

See also

Examples

# return summary organisation data for all organisations with 'Woodseats' in their name
getODS(Name="Woodseats")

#>                                 Name  OrgId   Status OrgRecordClass PostCode
#> 1 WELL WOODSEATS - CHESTERFIELD ROAD  FA460   Active            RC1   S8 0SE
#> 2       WELL WOODSEATS - COBNAR ROAD  FJX96   Active            RC1   S8 8QA
#> 3              WOODSEATS DENTAL CARE V05400   Active            RC1   S8 0GU
#> 4           WOODSEATS MEDICAL CENTRE  03NJW   Active            RC2   S8 0SH
#> 5           WOODSEATS MEDICAL CENTRE  5EPJW Inactive            RC2   S8 8QB
#> 6           WOODSEATS MEDICAL CENTRE  5EQJW Inactive            RC2   S8 8QB
#> 7           WOODSEATS MEDICAL CENTRE  5N4JW   Active            RC2   S8 8QB
#> 8           WOODSEATS MEDICAL CENTRE C88041   Active            RC1   S8 0SH
#> 9           WOODSEATS MEDICAL CENTRE  TAH96   Active            RC2   S8 8QB
#>   LastChangeDate PrimaryRoleId            PrimaryRoleDescription
#> 1     2016-03-15         RO182                          PHARMACY
#> 2     2015-08-19         RO182                          PHARMACY
#> 3     2016-07-07         RO110           GENERAL DENTAL PRACTICE
#> 4     2017-08-31          RO99 CLINICAL COMMISSIONING GROUP SITE
#> 5     2013-05-08         RO180           PRIMARY CARE TRUST SITE
#> 6     2013-05-08         RO180           PRIMARY CARE TRUST SITE
#> 7     2013-05-08         RO180           PRIMARY CARE TRUST SITE
#> 8     2018-01-10         RO177           PRESCRIBING COST CENTRE
#> 9     2013-05-08         RO108                   CARE TRUST SITE
#>                                                                 OrgLink
#> 1  https://directory.spineservices.nhs.uk/ORD/2-0-0/organisations/FA460
#> 2  https://directory.spineservices.nhs.uk/ORD/2-0-0/organisations/FJX96
#> 3 https://directory.spineservices.nhs.uk/ORD/2-0-0/organisations/V05400
#> 4  https://directory.spineservices.nhs.uk/ORD/2-0-0/organisations/03NJW
#> 5  https://directory.spineservices.nhs.uk/ORD/2-0-0/organisations/5EPJW
#> 6  https://directory.spineservices.nhs.uk/ORD/2-0-0/organisations/5EQJW
#> 7  https://directory.spineservices.nhs.uk/ORD/2-0-0/organisations/5N4JW
#> 8 https://directory.spineservices.nhs.uk/ORD/2-0-0/organisations/C88041
#> 9  https://directory.spineservices.nhs.uk/ORD/2-0-0/organisations/TAH96

# return summary organisation data for all organisations
# with 'Woodseats Medical Centre' in their name - two options to handle spaces:
getODS(Name="Woodseats_Medical_Centre")

#>                       Name  OrgId   Status OrgRecordClass PostCode
#> 1 WOODSEATS MEDICAL CENTRE  03NJW   Active            RC2   S8 0SH
#> 2 WOODSEATS MEDICAL CENTRE  5EPJW Inactive            RC2   S8 8QB
#> 3 WOODSEATS MEDICAL CENTRE  5EQJW Inactive            RC2   S8 8QB
#> 4 WOODSEATS MEDICAL CENTRE  5N4JW   Active            RC2   S8 8QB
#> 5 WOODSEATS MEDICAL CENTRE C88041   Active            RC1   S8 0SH
#> 6 WOODSEATS MEDICAL CENTRE  TAH96   Active            RC2   S8 8QB
#>   LastChangeDate PrimaryRoleId            PrimaryRoleDescription
#> 1     2017-08-31          RO99 CLINICAL COMMISSIONING GROUP SITE
#> 2     2013-05-08         RO180           PRIMARY CARE TRUST SITE
#> 3     2013-05-08         RO180           PRIMARY CARE TRUST SITE
#> 4     2013-05-08         RO180           PRIMARY CARE TRUST SITE
#> 5     2018-01-10         RO177           PRESCRIBING COST CENTRE
#> 6     2013-05-08         RO108                   CARE TRUST SITE
#>                                                                 OrgLink
#> 1  https://directory.spineservices.nhs.uk/ORD/2-0-0/organisations/03NJW
#> 2  https://directory.spineservices.nhs.uk/ORD/2-0-0/organisations/5EPJW
#> 3  https://directory.spineservices.nhs.uk/ORD/2-0-0/organisations/5EQJW
#> 4  https://directory.spineservices.nhs.uk/ORD/2-0-0/organisations/5N4JW
#> 5 https://directory.spineservices.nhs.uk/ORD/2-0-0/organisations/C88041
#> 6  https://directory.spineservices.nhs.uk/ORD/2-0-0/organisations/TAH96

getODS(Name="Woodseats Medical Centre")

#>                       Name  OrgId   Status OrgRecordClass PostCode
#> 1 WOODSEATS MEDICAL CENTRE  03NJW   Active            RC2   S8 0SH
#> 2 WOODSEATS MEDICAL CENTRE  5EPJW Inactive            RC2   S8 8QB
#> 3 WOODSEATS MEDICAL CENTRE  5EQJW Inactive            RC2   S8 8QB
#> 4 WOODSEATS MEDICAL CENTRE  5N4JW   Active            RC2   S8 8QB
#> 5 WOODSEATS MEDICAL CENTRE C88041   Active            RC1   S8 0SH
#> 6 WOODSEATS MEDICAL CENTRE  TAH96   Active            RC2   S8 8QB
#>   LastChangeDate PrimaryRoleId            PrimaryRoleDescription
#> 1     2017-08-31          RO99 CLINICAL COMMISSIONING GROUP SITE
#> 2     2013-05-08         RO180           PRIMARY CARE TRUST SITE
#> 3     2013-05-08         RO180           PRIMARY CARE TRUST SITE
#> 4     2013-05-08         RO180           PRIMARY CARE TRUST SITE
#> 5     2018-01-10         RO177           PRESCRIBING COST CENTRE
#> 6     2013-05-08         RO108                   CARE TRUST SITE
#>                                                                 OrgLink
#> 1  https://directory.spineservices.nhs.uk/ORD/2-0-0/organisations/03NJW
#> 2  https://directory.spineservices.nhs.uk/ORD/2-0-0/organisations/5EPJW
#> 3  https://directory.spineservices.nhs.uk/ORD/2-0-0/organisations/5EQJW
#> 4  https://directory.spineservices.nhs.uk/ORD/2-0-0/organisations/5N4JW
#> 5 https://directory.spineservices.nhs.uk/ORD/2-0-0/organisations/C88041
#> 6  https://directory.spineservices.nhs.uk/ORD/2-0-0/organisations/TAH96

# return summary organisation data for all currently active GP practices:
# commented out as takes too long to run with package build
# getODS(Status="Active", PrimaryRoleId = "RO177", NonPrimaryRoleId = "RO76")