getODSfull

Extracts full ODS data for a single organisation from the NHS Digital ODS ORD API into a list of length 1.

getODSfull(ODSCode)

Arguments

ODSCode
The organisation to return details for; quoted string; no default

Value

returns a list of length 1 containing the full details for the Organisation including: Name, Legal and Operational Start and End Dates, Status, Last Change Date, Organisation Record Class, Address, Postcode, Contact Information, Roles and Relationships to other Organisations

Notes

View the NHS Digital ODS API Implementation Guide at https://developer.nhs.uk/library/identifiers/ods-ord-api-implementation-guide/

View the NHS Digital ODS API Suite at https://directory.spineservices.nhs.uk/ODSAPISuite

See also

Examples

# return full Organisation data for RRF12
getODSfull(ODSCode="RRF12")

#> $Organisation
#> $Organisation$Name
#> [1] "PLATT BRIDGE CLINIC"
#> 
#> $Organisation$Date
#>          Type      Start
#> 1 Operational 2001-04-01
#> 
#> $Organisation$OrgId
#> $Organisation$OrgId$root
#> [1] "2.16.840.1.113883.2.1.3.2.4.18.48"
#> 
#> $Organisation$OrgId$assigningAuthorityName
#> [1] "HSCIC"
#> 
#> $Organisation$OrgId$extension
#> [1] "RRF12"
#> 
#> 
#> $Organisation$Status
#> [1] "Active"
#> 
#> $Organisation$LastChangeDate
#> [1] "2013-05-08"
#> 
#> $Organisation$orgRecordClass
#> [1] "RC2"
#> 
#> $Organisation$GeoLoc
#> $Organisation$GeoLoc$Location
#> $Organisation$GeoLoc$Location$AddrLn1
#> [1] "1 RIVINGTON AVENUE"
#> 
#> $Organisation$GeoLoc$Location$AddrLn2
#> [1] "PLATT BRIDGE"
#> 
#> $Organisation$GeoLoc$Location$Town
#> [1] "WIGAN"
#> 
#> $Organisation$GeoLoc$Location$County
#> [1] "LANCASHIRE"
#> 
#> $Organisation$GeoLoc$Location$PostCode
#> [1] "WN2 5NG"
#> 
#> $Organisation$GeoLoc$Location$Country
#> [1] "ENGLAND"
#> 
#> 
#> 
#> $Organisation$Roles
#> $Organisation$Roles$Role
#>      id uniqueRoleId primaryRole                    Date Status
#> 1 RO198          117        TRUE Operational, 2001-04-01 Active
#> 
#> 
#> $Organisation$Rels
#> $Organisation$Rels$Rel
#>                      Date Status                 Target.OrgId.root
#> 1 Operational, 2001-04-01 Active 2.16.840.1.113883.2.1.3.2.4.18.48
#>   Target.OrgId.assigningAuthorityName Target.OrgId.extension
#> 1                               HSCIC                    RRF
#>   Target.PrimaryRoleId.id Target.PrimaryRoleId.uniqueRoleId  id uniqueRelId
#> 1                   RO197                             61853 RE6       57641
#> 
#> 
#> $Organisation$Succs
#> $Organisation$Succs$Succ
#>   uniqueSuccId              Date        Type                 Target.OrgId.root
#> 1        14115 Legal, 2001-04-01 Predecessor 2.16.840.1.113883.2.1.3.2.4.18.48
#>   Target.OrgId.assigningAuthorityName Target.OrgId.extension
#> 1                               HSCIC                  RJY12
#>   Target.PrimaryRoleId.id Target.PrimaryRoleId.uniqueRoleId
#> 1                   RO198                             98719
#> 
#> 
#>